box: rewrite the comparator selection code
The previous commit introduced a lot of branching in the comparison
selection functions. The code may be simplified by adding more
wrappers with increasing number of template parameters.

This commit introduces another approach: using precalculated table
of slowpath, sequential and functional comparators. This gets rid
of the lot of branches, but adds a lot of templates, so the use of
this is change is to be discussed. Also it's limited: only bool
flags specialization is supported.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring