raft: speed up synchronous transaction queue processing
This patch optimizes the process of collecting ACKs from replicas for
synchronous transactions. Before this patch, this was not done
optimally, resulting in a possible situation where it was necessary to
go through the entire queue again each time when receiving the next ACK
from the replica. This was especially noticeable in the case of a large
number parallel synchronous requests. In this case, the performance was
about 8000 RPS on average.

Closes #9917

NO_DOC=performance improvement
NO_TEST=performance improvement