perf: `1mops_write --sync` does not count confirm
There was a problem in that when calculating the number of completed
operations, the test looked at `box.info.lsn`, which is bumped not only by
replaces in space, but also by confirms written to wal (in case of
synchronous replication). As a result, with the `--sync` flag set, the
number of operations was 2 times greater than the number of actual
replaces, and as a result, the rps it showed were 2 times greater than the
real ones.

Closes #11404
Closes #11407

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=internal